split out ctx and convert
